Updated some formatting/cleaned up some code
