+++ app/src/main/java/com/app/WeOut/EventHomeFeedDetailsActivity.java
@@ -53,7 +53,11 @@ public class EventHomeFeedDetailsActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_event_home_feed_details);
+
+        //create Firestore Database Reference
         this.dbReference = FirebaseFirestore.getInstance();
+
+        //Assign Views Based on Ids
         this.eventTitle = findViewById(R.id.eventHomeFeedClickedTitle);
         this.eventLocation = findViewById(R.id.eventHomeFeedClickedLocation);
         this.eventDate = findViewById(R.id.eventHomeFeedClickedDate);
@@ -62,6 +66,8 @@ public class EventHomeFeedDetailsActivity extends AppCompatActivity {
         this.totalAcceptedInvitees = findViewById(R.id.eventHomeFeedClickedAcceptedIntendees);
         this.eventDescription = findViewById(R.id.eventHomeFeedClickedDescription);
         this.eventOrganizer = findViewById(R.id.eventHomeFeedClickedOrganizer);
+
+        //Buttons on screen for closing and deleting an event
         this.closeScreen = findViewById(R.id.closeEventDetailButton);
         this.deleteEvent = findViewById(R.id.deleteEventButton);
 
@@ -72,6 +78,7 @@ public class EventHomeFeedDetailsActivity extends AppCompatActivity {
 
         final List<EventDetailsInvitee> inviteesForEvent = new ArrayList<>();
 
+        //we create a list of EventDetailsInvitee objects containing userName, first, last name and that they are attending the event
         for(Entry<String, String> invitee : eventDetailFromFeed.getAttendingMap().entrySet()) {
             String userName = invitee.getKey();
             String fullName = invitee.getValue();
@@ -82,16 +89,17 @@ public class EventHomeFeedDetailsActivity extends AppCompatActivity {
 
         this.totalAcceptedInvitees.setText(String.valueOf(attendingCount) + " yes");
 
+        //we create a list of EventDetailsInvitee objects containing userName, first, last name and that they are not attending the event
         for(Entry<String, String> invitee : eventDetailFromFeed.getInvitedMap().entrySet()) {
             String userName = invitee.getKey();
             String fullName = invitee.getValue();
             inviteesForEvent.add(new EventDetailsInvitee(userName, fullName.split(" ")[0], fullName.split(" ")[1], false));
         }
 
+        //calculate total invitees
         this.totalInvitees.setText(String.valueOf(eventDetailFromFeed.getInvitedMap().size() + attendingCount) + " guests");
 
-        this.myAdapter = new EventHomeFeedClickedInviteesRecyclerViewAdapter(inviteesForEvent);
-
+        //set text for information relating to an Event such as title, location, date
         this.eventTitle.setText(eventDetailFromFeed.getTitle());
 
         this.eventLocation.setText(eventDetailFromFeed.getLocation());
@@ -101,6 +109,9 @@ public class EventHomeFeedDetailsActivity extends AppCompatActivity {
         this.eventOrganizer.setText(eventDetailFromFeed.getOrganizer());
 
         this.eventDescription.setText(eventDetailFromFeed.getDescription());
+
+        //create a Custom Adapter for Our RecyclerView listing invitees
+        this.myAdapter = new EventHomeFeedClickedInviteesRecyclerViewAdapter(inviteesForEvent);
 
         this.eventGuests.setAdapter(this.myAdapter);
 
@@ -115,132 +126,51 @@ public class EventHomeFeedDetailsActivity extends AppCompatActivity {
             @Override
             public void onClick(final View view) {
                 //get Current Username of device
-                String userName = FirebaseAuth.getInstance().getCurrentUser().getEmail()
-                    .split("@weout.com")[0];
-
-                WriteBatch batchDeleteEvent = dbReference.batch();
+                String userName = Utilities.getCurrentUsername();
 
                 //if we are the organizer for the event and delete, need to remove from the rest of
                 //the invitees
                 if (userName.equals(eventDetailFromFeed.getOrganizer())) {
+                    try {
+                        Utilities.deleteEventAsOrganizer(eventWithIDDetailFromFeed, view,
+                            getApplicationContext());
+                        // Wait a bit to finish this activity
+                        final Handler handler = new Handler();
+                        handler.postDelayed(new Runnable() {
+                            @Override
+                            public void run() {
+                                // Do something after 5s = 5000ms
+                                finish();
+                            }
+                        }, 1000);
+                    } catch (IllegalStateException e) {
 
-
-                    //delete 'attendingMap' field from document with corresponding event ID
-                    HashMap<String, Object> deleteAttendingInvitees = new HashMap<>();
-                    deleteAttendingInvitees.put("attendingMap", FieldValue.delete());
-
-                    DocumentReference currentEventAttendingRef = dbReference.collection("events")
-                        .document(eventWithIDDetailFromFeed.getEventID());
-                    batchDeleteEvent.update(currentEventAttendingRef, deleteAttendingInvitees);
-
-                    //delete 'invitedMap' field from document with corresponding event ID
-                    HashMap<String, Object> deleteInvitedGuests = new HashMap<>();
-                    deleteAttendingInvitees.put("invitedMap", FieldValue.delete());
-
-                    DocumentReference currentEventInvitedRef = dbReference.collection("events")
-                        .document(eventWithIDDetailFromFeed.getEventID());
-                    batchDeleteEvent.update(currentEventInvitedRef, deleteInvitedGuests);
-
-                    //now delete the event id document
-                    DocumentReference currentEventRef = dbReference.collection("events")
-                        .document(eventWithIDDetailFromFeed.getEventID());
-                    batchDeleteEvent.delete(currentEventRef);
-
-                    //now we need to delete the event from each users invited and accepted document
-                    HashMap<String, Object> deleteInviteInUserInvites = new HashMap<>();
-                    deleteInviteInUserInvites.put(eventWithIDDetailFromFeed.getEventID(), FieldValue.delete());
-
-                    //traverse each invited user and delete from their event invites
-                    for(String username : eventDetailFromFeed.getInvitedMap().keySet()) {
-
-                            batchDeleteEvent.update(dbReference.collection("users")
-                                .document(username).collection("events")
-                                .document("invited"), deleteInviteInUserInvites);
                     }
 
-                    //traverse each invited user and delete from their event home feed
-                    for(String username : eventDetailFromFeed.getAttendingMap().keySet()) {
-
-                        batchDeleteEvent.update(dbReference.collection("users")
-                            .document(username).collection("events")
-                            .document("accepted"), deleteInviteInUserInvites);
-                    }
-
-                    batchDeleteEvent.commit().addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            Log.d(TAG, "Successfully Deleted Event: " + eventWithIDDetailFromFeed.getEventID());
-                            Utilities.displaySnackBar(view, getApplicationContext(), "Event Deleted", R.color.lightBlue);
-                            // Wait a bit to finish this activity
-                            final Handler handler = new Handler();
-                            handler.postDelayed(new Runnable() {
-                                @Override
-                                public void run() {
-                                    // Do something after 5s = 5000ms
-                                    finish();
-                                }
-                            }, 1500);
-                        }
-                    }).addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Log.d(TAG, "Unable to delete Event: " + eventWithIDDetailFromFeed.getEventID());
-                            Utilities.displaySnackBar(view, getApplicationContext(), "Error: Can't delete Event at this time", R.color.lightBlue);
-                        }
-                    });
 
                 }
                 //we need to delete the event from the user's events collection under the accepted document and we need to delete the event
                 //from the event's attendingMap
                 else {
+                    try {
+                        Utilities.deleteEventNonOrganizer(eventWithIDDetailFromFeed, view,
+                            getApplicationContext());
+                        // Wait a bit to finish this activity
+                        final Handler handler = new Handler();
+                        handler.postDelayed(new Runnable() {
+                            @Override
+                            public void run() {
+                                // Do something after 5s = 5000ms
+                                finish();
+                            }
+                        }, 1000);
+                    } catch (IllegalStateException e) {
 
-                    //create a reference to the user's events collection under the accepted document
-                    DocumentReference dbUserAcceptedRef = dbReference.collection("users")
-                        .document(userName).collection("events").document("accepted");
+                    }
 
-                    //delete the eventID from the user's accepted document
-                    HashMap<String, Object> deleteAcceptedEvent = new HashMap<>();
-                    deleteAcceptedEvent.put(eventWithIDDetailFromFeed.getEventID(), FieldValue.delete());
-
-                    batchDeleteEvent.update(dbUserAcceptedRef, deleteAcceptedEvent);
-
-                    //create a reference to the event's attendingMap Field
-                    DocumentReference dbEventAttending = dbReference.collection("events")
-                        .document(eventWithIDDetailFromFeed.getEventID());
-
-                    //delete the username key from the attendingMap Map
-                    HashMap<String, Object> deleteUsernameFromAttendingMap = new HashMap<>();
-                    deleteUsernameFromAttendingMap.put("attendingMap." + userName, FieldValue.delete());
-
-                    batchDeleteEvent.update(dbEventAttending,deleteUsernameFromAttendingMap);
-
-                    batchDeleteEvent.commit().addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            Log.d(TAG, "Successfully Deleted Event: " + eventWithIDDetailFromFeed.getEventID());
-                            Utilities.displaySnackBar(view, getApplicationContext(), "Event Deleted", R.color.lightBlue);
-                            // Wait a bit to finish this activity
-                            final Handler handler = new Handler();
-                            handler.postDelayed(new Runnable() {
-                                @Override
-                                public void run() {
-                                    // Do something after 5s = 5000ms
-                                    finish();
-                                }
-                            }, 1500);
-
-                        }
-                    }).addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Log.d(TAG, "Unable to delete Event: " + eventWithIDDetailFromFeed.getEventID());
-                            Utilities.displaySnackBar(view, getApplicationContext(), "Error: Can't delete Event at this time", R.color.lightBlue);
-                        }
-                    });
                 }
-
             }
         });
-
     }
 }
+
