+++ app/src/main/res/layout/fragment_main_activity_add_event.xml
@@ -135,7 +135,6 @@
                         android:layout_column="0"
                         android:layout_columnWeight="1"
                         android:text="@string/eventCreationSetDateButton"
-                        android:background="@drawable/bluebutton"
                         android:textAllCaps="false"
                         />
 
@@ -146,7 +145,6 @@
                         android:layout_column="1"
                         android:layout_columnWeight="1"
                         android:text="@string/eventCreationSetTimeButton"
-                        android:background="@drawable/bluebutton"
                         android:textAllCaps="false"
                         />
 
@@ -226,7 +224,6 @@
                     android:layout_marginVertical="5dp"
                     android:text="@string/eventCreationInviteFriendsButton"
                     android:textAllCaps="false"
-                    android:background="@drawable/bluebutton"
                     />
 
             </RelativeLayout>
